#35 Added ShopItemUnlocker based refactoring.
Skin and Soundtrack have now been refactored even more. The responsibility about keeping an HashMap is now taken away and the responsibility of the ShopItemUnlocker. Skin and Soundtrack are now abstract classes!
Besides, all fields that were in Skin and Soundtrack have been moved to ShopItemLoader, as he is responsible for loading skins and soundtracks.

<feature,organisation,documentation